// src/constants/recipe.interface.ts
export interface Recipe {
  inspired_by_url?: string;
  thumbnail_alt_text: string;
  total_time_minutes?: number;
  approved_at: number;
  instructions: {
    position: number;
    display_text: string;
    start_time?: number;
    appliance?: string;
    end_time?: number;
    temperature?: number;
    id: number;
  }[];
  id: number;
  brand_id?: string;
  num_servings: number;
  video_id?: string;
  is_one_top: boolean;
  servings_noun_plural: string;
  renditions: any[];
  video_ad_content?: string;
  slug: string;
  compilations: any[];
  description: string;
  thumbnail_url: string;
  brand?: string;
  name: string;
  video_url?: string;
  credits: { name: string; image_url?: string; type: string }[];
  promotion: string;
  nutrition_visibility: string;
  buzz_id?: string;
  tips_and_ratings_enabled: boolean;
  draft_status: string;
  show_id: number;
  prep_time_minutes?: number;
  tags: { name: string; id: number; display_name: string; type: string }[];
  aspect_ratio: string;
  facebook_posts: any[];
  language: string;
  user_ratings: { count_positive: number; score?: number; count_negative: number };
  servings_noun_singular: string;
  cook_time_minutes?: number;
  show: { name: string; id: number };
  updated_at: number;
  beauty_url?: string;
  canonical_id: string;
  topics: { slug: string; name: string }[];
  seo_title: string;
  original_video_url?: string;
  country: string;
  keywords: string;
  nutrition?: {
    carbohydrates: number;
    fiber: number;
    updated_at?: string;
    protein: number;
    fat: number;
    calories: number;
    sugar: number;
  };
  created_at: number;
  is_shoppable: boolean;
  total_time_tier?: { tier: string; display_tier: string };
  yields: string;
}